lint: fix golangci findings; run the action's v2-capable major
golangci-lint-action v6 cannot run a golangci v2 config (the lint job has
been failing at setup), so bump to the action's current major. Fixing the
findings the working linter reports: unchecked Sscanf/Close/Remove returns
(errcheck), an ineffectual cont assignment in the captured-loop lowering
(ineffassign), a tagged switch in emitBox and De Morgan in the shim
signature scanner (staticcheck), and two redundant uint32 conversions in
the IL emitter (unconvert). No behavior change; unit tests and fixture
outputs identical.

+++ internal/emit/il.go
@@ -100,7 +100,7 @@ func translateMethod(m *goir.Method, tok tokenSet, localSigTok uint32) []byte {
 			b.buf = binary.LittleEndian.AppendUint32(b.buf, math.Float32bits(float32(op.Float)))
 		case goir.OpLdStr:
 			b.u8(0x72)
-			b.u32(0x70000000 | uint32(tok.us[op.Str]))
+			b.u32(0x70000000 | tok.us[op.Str])
 		case goir.OpLdLoc:
 			localOp(0x11, 0x0C, op.Local) // ldloc.s / ldloc
 		case goir.OpStLoc:
@@ -236,7 +236,7 @@ func translateMethod(m *goir.Method, tok tokenSet, localSigTok uint32) []byte {
 				b.u32(tokStrFromLitBytes)
 			} else {
 				b.u8(0x72) // ldstr (System.String)
-				b.u32(0x70000000 | uint32(tok.us[op.Str]))
+				b.u32(0x70000000 | tok.us[op.Str])
 				b.u8(0x28) // call GoStrings.FromLiteral -> GoString
 				b.u32(tokStrFromLit)
 			}
